Skip check for duplicates when GUID is nil
If a user do not set a GUID for more than one episode, it will be nil,
and the following warning will appear:

Warning: Multiple episodes are using the same GUID: `{nil=>[nil, nil]}`.

However, this is misleading, since the GUID will be replaced by a random
one and thus the episodes will have different GUIDs.